+++ CutTheRopeWP7/ctr_commons/CtrRenderer.cs
@@ -131,7 +131,7 @@ namespace ctr_wp7.ctr_commons
                         {
                             num5 = 1L;
                         }
-                        int num6 = (int)(1000000000L * (long)num4 / num5);
+                        int num6 = (int)(1000000000L * num4 / num5);
                         playedTicks += DELTA_NANOS;
                         if (num2 - playedTicks < DELTA_NANOS_THRES)
                         {
@@ -264,8 +264,8 @@ namespace ctr_wp7.ctr_commons
         // Token: 0x060006FF RID: 1791 RVA: 0x00038944 File Offset: 0x00036B44
         public static void Java_com_zeptolab_ctr_CtrRenderer_nativeResize(int width, int height, bool isLowMem)
         {
-            REAL_SCREEN_WIDTH = (float)width;
-            REAL_SCREEN_HEIGHT = (float)height;
+            REAL_SCREEN_WIDTH = width;
+            REAL_SCREEN_HEIGHT = height;
             SCREEN_RATIO = REAL_SCREEN_HEIGHT / REAL_SCREEN_WIDTH;
             IS_WVGA = width > 500 || height > 500;
             IS_QVGA = width < 280 || height < 280;
@@ -280,8 +280,8 @@ namespace ctr_wp7.ctr_commons
                 VIEW_SCREEN_HEIGHT = REAL_SCREEN_HEIGHT;
                 VIEW_SCREEN_WIDTH = SCREEN_WIDTH * REAL_SCREEN_HEIGHT / SCREEN_HEIGHT;
             }
-            VIEW_OFFSET_X = ((float)width - VIEW_SCREEN_WIDTH) / 2f;
-            VIEW_OFFSET_Y = ((float)height - VIEW_SCREEN_HEIGHT) / 2f;
+            VIEW_OFFSET_X = (width - VIEW_SCREEN_WIDTH) / 2f;
+            VIEW_OFFSET_Y = (height - VIEW_SCREEN_HEIGHT) / 2f;
             SCREEN_HEIGHT_EXPANDED = SCREEN_HEIGHT * REAL_SCREEN_HEIGHT / VIEW_SCREEN_HEIGHT;
             SCREEN_WIDTH_EXPANDED = SCREEN_WIDTH * REAL_SCREEN_WIDTH / VIEW_SCREEN_WIDTH;
             SCREEN_OFFSET_Y = (SCREEN_HEIGHT_EXPANDED - SCREEN_HEIGHT) / 2f;
@@ -290,7 +290,7 @@ namespace ctr_wp7.ctr_commons
             SCREEN_BG_SCALE_X = SCREEN_WIDTH_EXPANDED / SCREEN_WIDTH;
             if (IS_WVGA)
             {
-                SCREEN_WIDE_BG_SCALE_Y = (float)((double)SCREEN_HEIGHT_EXPANDED * 1.5 / 800.0);
+                SCREEN_WIDE_BG_SCALE_Y = (float)(SCREEN_HEIGHT_EXPANDED * 1.5 / 800.0);
                 SCREEN_WIDE_BG_SCALE_X = SCREEN_BG_SCALE_X;
                 return;
             }
@@ -494,7 +494,7 @@ namespace ctr_wp7.ctr_commons
         private static long DELTA_NANOS = 18181818L;
 
         // Token: 0x04000C97 RID: 3223
-        private static long DELTA_NANOS_THRES = (long)((double)DELTA_NANOS * 0.35);
+        private static long DELTA_NANOS_THRES = (long)(DELTA_NANOS * 0.35);
 
         // Token: 0x04000C98 RID: 3224
         private static bool DRAW_NOTHING;
